requirements.txt
streamlit 
crewai
crewai[tools]
python-dotenv
openai
langchain_openai
pydub
python-docx
groq